# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  serialize.ClusterInfo:
    properties:
      errors_count_threshold:
        example: 0
        type: integer
      expiration_timeout:
        example: 0
        type: integer
      host:
        example: host
        type: string
      kill_by_memory_with_dump:
        example: true
        type: boolean
      kill_problem_processes:
        example: true
        type: boolean
      life_time_limit:
        example: 0
        type: integer
      lifetime_limit:
        example: 0
        type: integer
      load_balancing_mode:
        description: performance
        example: 0
        type: integer
      max_memory_size:
        example: 0
        type: integer
      max_memory_time_limit:
        example: 0
        type: integer
      name:
        example: Новый кластер
        type: string
      port:
        example: 1542
        type: integer
      security_level:
        example: 0
        type: integer
      session_fault_tolerance_level:
        example: 0
        type: integer
      uuid:
        example: 6d6958e1-a96c-4999-a995-698a0298161e
        type: string
    type: object
...
  serialize.UserInfo:
    properties:
      descr:
        example: дополнительное описание
        type: string
      name:
        example: admin
        type: string
      password:
        type: string
      password_auth_allowed:
        type: boolean
      sys_auth_allowed:
        type: boolean
      sys_user_name:
        type: string
    type: object
host: localhost:3001
info:
  contact:
    email: khorevaa@yandex.ru
    name: API Support
  description: KUBOdin Swagger UI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'KUBOdin: Remote Administration for 1S.Enterprise Application Servers'
  version: "1.0"
paths:
...
  /app/{app}/cluster/{cluster}/admins:
    get:
      consumes:
      - application/json
      description: получение списка администраторов кластера
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serialize.UserInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение списка администраторов кластера
      tags:
      - admins
      - clusters
    post:
      consumes:
      - application/json
      description: выполняет регистрацию нового администратор на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: user info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/serialize.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.UserInfo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: выполняет регистрацию нового администратор на кластере
      tags:
      - admins
      - clusters
  /app/{app}/cluster/{cluster}/admins/{admin}:
    delete:
      consumes:
      - application/json
      description: Удаление администратора агента на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: admin name
        in: path
        name: admin
        required: true
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Удаление администратора агента на кластере
      tags:
      - admins
      - clusters
  /app/{app}/clusters:
    get:
      consumes:
      - application/json
      description: получение списка кластеров на сервере 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serialize.ClusterInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение списка кластеров на сервере 1С Предприятие
      tags:
      - clusters
  /app/{app}/clusters/{cluster}:
    get:
      consumes:
      - application/json
      description: получение информации о кластере на сервере 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.ClusterInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение информации о кластере на сервере 1С Предприятие
      tags:
      - clusters
...
swagger: "2.0"